refactor: package name 소문자로 변경
