+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/domains/promiseinteraction/domain/PromiseInteraction.kt
@@ -1,4 +1,4 @@
-package com.depromeet.whatnow.domains.pictureComment.domain
+package com.depromeet.whatnow.domains.promiseinteraction.domain
 
 import com.depromeet.whatnow.common.BaseTimeEntity
 import javax.persistence.Column
@@ -9,16 +9,18 @@ import javax.persistence.Id
 import javax.persistence.Table
 
 @Entity
-@Table(name = "tbl_picture_comment")
-class PictureComment(
-    var pictureId: Long,
+@Table(name = "tbl_promise_interaction")
+class PromiseInteraction(
+    var promiseId: Long,
+
+    var interactionId: Long,
 
     var userId: Long,
 
-    var comment: String,
+    var count: Long,
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "picture_comment_id")
+    @Column(name = "promise_interaction_id")
     val id: Long? = null,
 ) : BaseTimeEntity()
